feat(notifications): renaming code confirmation table to otp
